requirements.txt
Django==1.7.1
Markdown==2.1.1
PyYAML==3.11
argh==0.23.2
argparse==1.2.1
coverage==3.6
django-nose==1.2
djangorestframework>=2.3.8
django-filter==0.11.0
jsonschema==2.5
nose==1.3.0
mock==1.0.1
ordereddict==1.1
docutils>=0.11
